Корректировка самой первой версии проекта
