+++ client/src/main/java/ru/paramonov/project/client/Application.java
@@ -1,0 +1,9 @@
+package ru.paramonov.project.client;
+
+
+public class Application
+{
+    public static void main( String[] args ){
+
+    }
+}
