+++ client/src/main/java/ru/paramonov/project/client/common/LocalCard.java
@@ -1,0 +1,34 @@
+package ru.paramonov.project.client.common;
+
+import ru.paramonov.project.server.common.Card;
+
+public class LocalCard implements Card {
+    private final int cardNumber;
+    private final int personalIdentificationNumber;
+    private final long accountNumber;
+
+    public LocalCard(int cardNumber, long accountNumber, int personalIdentificationNumber) {
+        this.cardNumber = cardNumber;
+        this.accountNumber = accountNumber;
+        this.personalIdentificationNumber = personalIdentificationNumber;
+    }
+
+    @Override
+    public long getAccountNumber() {
+        return accountNumber;
+    }
+
+    @Override
+    public int getCardNumber() {
+        return cardNumber;
+    }
+
+    @Override
+    public String getBankName() {
+        return "MyBank";
+    }
+
+    public int getPersonalIdentificationNumber() {
+        return personalIdentificationNumber;
+    }
+}
