+++ server/src/main/java/ru/paramonov/project/server/Application.java
@@ -1,0 +1,9 @@
+package ru.paramonov.project.server;
+
+
+public class Application
+{
+    public static void main( String[] args ){
+
+    }
+}
