+++ server/src/main/java/ru/paramonov/project/server/common/Card.java
@@ -1,0 +1,12 @@
+package ru.paramonov.project.server.common;
+
+public interface Card {
+    int cardNumber = 0;
+    int personalIdentificationNumber = 0;
+    String bankName = null;
+
+    int getCardNumber();
+    long getAccountNumber();
+    int getPersonalIdentificationNumber();
+    String getBankName();
+}
