+++ server/src/main/java/ru/paramonov/project/server/common/Currency.java
@@ -1,0 +1,7 @@
+package ru.paramonov.project.server.common;
+
+public enum Currency {
+        RUB,
+        USD,
+        EUR
+    }
