+++ server/src/main/java/ru/paramonov/project/server/exception/PersonalIdentificatorNumberException.java
@@ -1,0 +1,8 @@
+package ru.paramonov.project.server.exception;
+
+public class PersonalIdentificatorNumberException extends Exception {
+    public String toString()
+    {
+        return "Incorrect Personal Identificator Number(PIN).";
+    }
+}
