--- src/main/java/ru/paramonov/project/Application.java
@@ -1,9 +1,0 @@
-package ru.paramonov.project;
-
-
-public class Application
-{
-    public static void main( String[] args ){
-
-    }
-}
