--- src/main/java/ru/paramonov/project/client/Banknote.java
@@ -1,4 +1,0 @@
-package ru.paramonov.project.client;
-
-public interface Banknote {
-}
