--- src/main/java/ru/paramonov/project/exceptions/PersonalIdentificatorNumberException.java
@@ -1,8 +1,0 @@
-package ru.paramonov.project.exceptions;
-
-public class PersonalIdentificatorNumberException extends Exception {
-    public String toString()
-    {
-        return "Incorrect Personal Identificator Number(PIN).";
-    }
-}
