--- src/main/java/ru/paramonov/project/server/Account.java
@@ -1,20 +1,0 @@
-package ru.paramonov.project.server;
-
-import lombok.Getter;
-import java.math.BigDecimal;
-import java.time.LocalDate;
-
-@Getter
-public class Account {
-    private final long accountNumber;
-    private final BigDecimal balance;
-    private final Currency accountCurrency;
-    private final LocalDate creationDate;
-
-    public Account(long accountNumber, BigDecimal balance, Currency accountCurrency, LocalDate creationDate) {
-        this.accountNumber = accountNumber;
-        this.balance = balance;
-        this.accountCurrency = accountCurrency;
-        this.creationDate = creationDate;
-    }
-}
