--- src/main/java/ru/paramonov/project/server/Currency.java
@@ -1,7 +1,0 @@
-package ru.paramonov.project.server;
-
-public enum Currency {
-        RUB,
-        USD,
-        EUR
-    }
