--- src/test/java/ru/paramonov/project/ApplicationTest.java
@@ -1,13 +1,0 @@
-package ru.paramonov.project;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import org.junit.jupiter.api.Test;
-
-public class ApplicationTest
-{
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
